requirements.txt
Django
djangorestframework
django-allauth
psycopg2-binary
python-decouple
dj-rest-auth
requests